+++ src/main/java/com/codefork/aoc2024/day17/Operand.java
@@ -1,11 +1,11 @@
 package com.codefork.aoc2024.day17;
 
 public abstract sealed class Operand permits Literal, Combo {
-    protected final int value;
+    protected final long value;
 
-    public Operand(int value) {
+    public Operand(long value) {
         this.value = value;
     }
 
-    public abstract int eval(Computer computer);
+    public abstract long eval(Computer computer);
 }
